Move ProcessFactory to the right namespace
